longarr phase 8a: alignMask uint32 truncation + huge-array probes
MemoryModel::allocate/free/reallocate at src/misc/memory_model.cpp:122/151/194-195
mask uint64 size with ~alignMask. alignMask was uint32_t, so ~alignMask zero-extends
to 0x00000000FFFFFFF0 when ANDed with the uint64 size — silently dropping the high
32 bits for any allocation >= 4 GB. A 4 GB+15 request became 16 bytes; the function
then took the shoe path with size=0, computed si = (0>>4)-1 = 0xFFFFFFFF, and
dereferenced chunks[0xFFFFFFFF] — a wild-address read that crashed the process.

Phase 1 widened the heap public API to uint64 but missed this field on both
MemoryModel and LinearChunkAllocator. Fix is a one-word widening of each;
the existing `(size + alignMask) & ~alignMask` lines pick up the wider type
automatically, and the `DAS_VERIFYF(s <= UINT32_MAX)` policy guard in
LinearChunkAllocator::allocate now fires correctly on >4 GB requests instead
of seeing a silently-truncated size.

Tests, all gated on DASLANG_HUGE_HEAP_TESTS=1:
 - tests-cpp/small/test_heap_64bit.cpp — new 4 GB-boundary test asserting
   bytesAllocated grows by >= 4 GB through PersistentHeapAllocator; existing
   5 GB test moved to persistent_heap (default LinearHeapAllocator is uint32-
   bounded by design and now panics with a clear message).
 - tests/long_array_table/test_huge_array_resize_index.das (5 GB array<uint8>)
 - tests/long_array_table/test_huge_array_iterate.das (2.2 GB, four iteration shapes)
 - tests/long_array_table/test_huge_array_push_emplace_clone.das (push past INT_MAX)
 - tests/long_array_table/test_huge_array_index_offset.das (~4.4 GB array<int>;
   exercises uint64 stride*idx address math)

All four daslang probes carry `options persistent_heap = true` (required for
>4 GB arrays) and an inline gate (`static_if (typeinfo sizeof(type<int?>) < 8)`
+ has_env_variable check) so they silent-skip in CI without the env var.

+++ include/daScript/misc/memory_model.h
@@ -302,7 +302,10 @@ namespace das {
         void setTrackAllocations ( bool on );
         __forceinline bool isTrackingAllocations() const { return trackAllocations; }
         CustomGrowFunction      customGrow;
-        uint32_t                alignMask;
+        // Mask must be uint64 — `(size + alignMask) & ~alignMask` in allocate/free/reallocate
+        // would otherwise zero-extend `~alignMask` from uint32 to uint64 as 0x00000000FFFFFFF0,
+        // silently truncating any allocation ≥ 4 GB to its low 32 bits.
+        uint64_t                alignMask;
         uint64_t                totalAllocated;
         uint64_t                maxAllocated;
         uint64_t                initialSize = 0;
@@ -411,7 +414,10 @@ namespace das {
         CustomGrowFunction  customGrow;
         uint64_t    unadjustedInitialSize = 0;
         uint64_t    initialSize = 0;
-        uint32_t    alignMask = 15;
+        // uint64 — see MemoryModel::alignMask. `~alignMask` must be uint64 so the
+        // `DAS_VERIFYF(s <= UINT32_MAX)` cap check in allocate() actually fires on
+        // >4 GB requests instead of seeing a silently-truncated low-32-bit size.
+        uint64_t    alignMask = 15;
         HeapChunk * chunk = nullptr;
     };
